requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
reportlab>=4.1.0
openpyxl>=3.1.2
python-dateutil>=2.9.0
Pillow>=10.2.0
plotly>=5.20.0
bcrypt>=4.1.2
